app/requirements.txt
numpy==1.26.3
pandas>=1.5.0
streamlit==1.30.0
tensorflow==2.15.0.post1
Pillow>=9.0.0
plotly>=5.0.0